Attempt to bypass max_input_vars
PHP imposes a limit on the size of an array to 1000 rows by default. Short of altering php.ini, selecting >1000 rows in the files list means saves aren't committed. To bypass this, the plugin now:

a) intercepts the submission process and serializes the form content
b) removes all input fields and adds this serialized data as the only field
c) submits the new form
d) recontructs the $_POST array so the plugin can continue as if it had received a full, regular payload

The only thing to watch now is that the max post payload (and/or suhosin equivalent) isn't exceeded.

Hat tip: https://stackoverflow.com/questions/30830447/workaround-for-max-input-vars-limit